pyproject.toml
[project]
name = "auto-cro-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cachetools>=6.2.6",
    "fastapi>=0.135.1",
    "mlflow>=3.10.1",
    "numpy>=2.4.3",
    "openrouter-client>=0.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.2",
    "torch>=2.10.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]
